Moved gif related functions into separate file
